Refactor into more files
