+++ apps/typegpu-docs/src/examples/simple/rotating-triangle-tiles/buffers.ts
@@ -1,0 +1,9 @@
+import * as d from 'typegpu/data';
+import { colors } from './geometry.ts';
+import { root } from './root.ts';
+
+const animationProgress = root.createUniform(d.f32);
+
+const shiftedColors = root.createReadonly(d.arrayOf(d.vec4f, 3), [...colors]);
+
+export { animationProgress, shiftedColors };
