+++ apps/typegpu-docs/src/examples/simple/rotating-triangle-tiles/root.ts
@@ -1,0 +1,3 @@
+import tgpu from 'typegpu';
+
+export const root = await tgpu.init();
